pyproject.toml
[project]
name = "vidiosummary"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.58.1",
    "python-dotenv>=1.0.1",
    "pytube>=15.0.0",
    "pytubefix>=8.8.2",
    "requests>=2.32.3",
    "scrapy-playwright>=0.0.42",
    "scrapy>=2.12.0",
    "selenium>=4.27.1",
    "tensorflow-hub>=0.16.1",
    "tensorflow>=2.18.0",
    "youtube-dl>=2021.12.17",
    "youtube-transcript-api>=0.6.3",
    "yt-dlp>=2024.12.13",
]
